requirements_runocr.txt
aiohttp>=3.10.10
fastapi>=0.115.4
numpy>=2.1.2
opencv-python>=4.10.0.84
Pillow>=11.0.0
python-dotenv>=1.0.1
python-multipart>=0.0.16
torch>=2.5.1
torchvision>=0.20.1
transformers>=4.46.1
uvicorn>=0.32.0
